backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
httpx==0.25.1
praw==7.7.1
tweepy==4.14.0
openai==1.3.7
scikit-learn==1.3.2
numpy==1.24.3
python-multipart==0.0.6
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4